fix typing error and move sql test data to a single file
